Add local storage mock to a file for reuse
