+++ packages/components/organisms/f-checkout/test-utils/local-storage/local-storage-mock.js
@@ -1,0 +1,19 @@
+const localStorageMock = (function localStorageMock () {
+    let store = {};
+    return {
+        getItem: function getItem (key) {
+            return store[key];
+        },
+        setItem: function setItem (key, value) {
+            store[key] = value.toString();
+        },
+        clear: function clear () {
+            store = {};
+        },
+        removeItem: function removeItem (key) {
+            delete store[key];
+        }
+    };
+}());
+
+export default localStorageMock;
